PA-3360: Moved MetadataProvider to metadata directory
