+++ Src/SwqlStudio/Metadata/SwisMetaDataProvider.cs
@@ -2,9 +2,8 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
-using SwqlStudio.Metadata;
 
-namespace SwqlStudio
+namespace SwqlStudio.Metadata
 {
     class SwisMetaDataProvider : IMetadataProvider
     {
