+++ Src/SwqlStudio/ServerList.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using SwqlStudio.Metadata;
 
 namespace SwqlStudio
 {
